move module-load config from /etc to /usr
As /etc is reserved for the admin and /usr is the one for defaults
shipped by packages. A admin can simply override this file by placing
one with the same name inside /etc/modules-load.d, so if they modified
the (previously) shipped one there it will stay active, so no
regression potential with upgrades here.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>