vm start/commandline: activate volumes before config_to_command()
With '-blockdev', it is necessary to activate the volumes to generate
the command line, because it can be necessary to check whether the
volume is a block device or a regular file.

Do not deactivate after commandline generation for 'qm showcmd',
because there can be concurrent operations acting on the volumes.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>