fix #5985: qmp client: increase timeout for {device, netdev, object}_{add, del} commands
In the bug report, the user mentioned that 7 seconds was enough. For
the HMP 'drive_add' command, the used timeout is 1 minute and for the
'drive_del' command, the used timeout is 10 minutes, because IO might
need to be finished. While something similar might be true for certain
objects/devices, there were no issues reported with the *_del
operations using the default timeout until now and the callers can
still use a higher timeout if they know the specific device/object
requires it.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>