blockdev: also set read-only flag on top throttle node
QEMU will try to set the write permission on a device even if it's
just the throttle filter that is not marked as read-only. This causes
VM start to fail when a drive with the Proxmox VE-specific 'ro' flag
is configured and also template backup. Properly set the flag also
for the top throttle node.

Fixes: 2ea50f8b ("blockdev: add helpers to generate blockdev commandline")
Reported-by: Aaron Lauterer <a.lauterer@proxmox.com>
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250718154503.179257-1-f.ebner@proxmox.com