tests: cfg2cmd: improve mocking File::stat
Stat-ing /dev/null worked, because the check in the blockdev module is
for block and character devices and then decides based on the
'media=cdrom' flag whether to use the host_cdrom or host_device
driver. But the result should actually be mocked to represent a block
device rather than a character device.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>