block job: mirror: always detach the target node upon error
For example, attempting to live-migrate a disk again after failure
would not work, because a node with the same name would still be
attached.

Mirroring the disk to a shared storage after VM live-migration failure
and then attempting VM live-migration again, would result in

> migration status error: failed - Error in migration completion: Input/output error

This is because for migration completion, all attached block devices
are flushed, but the NBD export does not exist on the target side
anymore.

Fixes: 1da91175 ("block job: add blockdev mirror")
Reported-by: Hannes Dürr <h.duerr@proxmox.com>
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Tested-by: Hannes Duerr <h.duerr@proxmox.com>
Link: https://lore.proxmox.com/20250728100702.38685-1-f.ebner@proxmox.com