qemu-server: exit delete early if we cannot find a snapshot
Die if the list of snapshots returned by `volume_snapshot_info` does
not contain the snapshot we are trying to delete.

Previously it was just assumed that the snapshot would be present,
leading to the entry in the list being autoviviefied by the following
lines of code. Hence, we tried to commit nonexistent snapshot states
here, as both `$parentsnap` and `$childsnap` would be undefined.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20250730103705.98313-3-s.sterz@proxmox.com
[FE: also include volid in error message]
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>