migration: only use nets-host-mtu for PVE 8 target if actually required
Commit 1736fbeb ("migration: preserve host_mtu for virtio-net
devices") addresses probblematic migration scenarios with VirtIO-net
devices which inherit the MTU from the bridge. In Proxmox VE 9, the
default behavior changed to inherit the MTU from the bridge, but for
Proxmox VE 8, only an explicit mtu=1 will do so and thus be
potentially problematic.

By not using the nets-host-mtu parameter for unproblematic migrations,
fricition is reduced, because backwards migrations to not-yet-upgraded
Proxmox VE 8 nodes will still work.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20250909091918.32254-4-f.ebner@proxmox.com